requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
